Change properties to have a pending value and report valid values.
Pending values are transferred to current values at mode set, allowing a
queue of values to accrue in preparation for the modeset operation while
still reporting valid 'current' values for applications.

The set of valid values is also available so applications can present
a reasonable list to the user during configuration.

I renumbered a bunch of requests to keep property requests sequential.