Add major/minor version as supported by client to GetExtensionVersionReq.
This sort-of breaks old clients. Behaviour to be assumed is that if nbytes is
0, major/minorVersion is set and specifies the version as supported by the
client.
If nbytes is non-zero, the request is trailed by the extension name (INAME)
and major/minorVersion is undefined. This is the behaviour of pre-MPX clients.

And then there may be clients who found that no other extension uses this
request and supplying a name wasn't actually necessary since it was XI anyway.
These clients will break. Tough luck. Read the man pages next time.