Define the Cursor datasize correctly
On 64 bit machines, without Cursor defined Xlib would allocate 64 bits
rather than 32 to any structs using Cursor. This led to data not
correctly being available on the wire hence the Xserver would do strange
things. We hence define Cursor to what it should be and make sure
we undefine it after we've finished to users of XIproto.h aren't affected

Fix-by: Peter Hutterer <peter.hutterer@who-t.net>
Signed-off-by: Benjamin Close <Benjamin.Close@clearchain.com>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>