Clarify semantics of clipping
Source clips are applied on top of the transformed and repeated
window, because that way you can clip out an interesting shape of for
example a repeated gradient, and because this makes the implementation
simple and efficient.

GraphicsExposures are never generated.

The hierarchy clip is ignored for source and mask pictures. This is
because GraphicsExposures are not generated anyway, and because
implementing it correctly in the presense of filters and
transformations would be either inefficient or very complex.

Also make it explicit that the dither argument is not used.