Change xf86dga.h to only conditionally include server's header file.
The recent re-addition of the xf86dga.h file introduced an
incompatibility with a recent, but unreleased version of libXxf86dga
which also wanted to install a file of the same name. Instead, we
allow the proto package to own this file, and include the library's
version (now to be named Xxf86dga.h) unless we're building an old
server (indicated by _XF86DGA_SERVER_).