Add sourceid to RawEvents (#34420)
RawEvents in XI2 do not provide the source ID. The libXi headers however do
and it is currently always 0. Given that the sourceid may be useful for
some clients, send it down the wire.

This has no effect on the wire size of the struct, we can re-use a pad byte
here.

X.Org Bug 34420 <http://bugs.freedesktop.org/show_bug.cgi?id=34420>

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Reviewed-by: Daniel Stone <daniel@fooishbar.org>