specs: fix multi licensed copyright notice and license text
For documentation having multiple licenses, the copyright and legalnotice
elements sequence cannot instantiated multiple times.
The copyright notice and license text are therefore coded inside a legalnotice
element. The role attribute on the paragraph is used to allow styling of the
copyright notice text which should not be italicized.

To view the change in the html format, obtain xorg-sgml-doctools
at version 1.10.1 or later and rebuild the html version of the spec.

Signed-off-by: Gaetan Nadon <memsize@videotron.ca>