glproto: Document GLX_ARB_create_context and GLX_ARB_create_context_profile protocol
Also add the struct names (in addition to the typedef names).  Most of
the other protocol structures have this, but some lack it.

v2: Add a few missing Req suffixes.  Rename n0, n1, and n2 fields of
GLXSetClientInfoARB and GLXSetClientInfo2ARB to numVersions,
numGLExtensionBytes and numGLXExtensionBytes, respectively.  Since the
names of some structure fields change, bump the version to 1.4.15.

Signed-off-by: Ian Romanick <ian.d.romanick@intel.com>
Reviewed-by: Dave Airlie <airlied@redhat.com>