Xvproto: fix xvEncodingInfo to match actual wire protocol
The padding is *before* the rate field, so the rate is placed on a 32bit
boundary.  This change adds explicit padding between height and rate,
and removes extraneous padding after the rate field, which the server
never sent and xlib never read.

This changes sizeof(xvEncodingInfo).  Hopefully that's not a big deal as
clients only see the Xlib structure XvEncodingInfo.

Debian bug#657215

Reported-by: Kevin Ryde <user42@zip.com.au>
Signed-off-by: Julien Cristau <jcristau@debian.org>
Reviewed-by: Daniel Martin <consume.noise@gmail.com>
Reviewed-by: Alan Coopersmith <alan.coopersmith@oracle.com>
Signed-off-by: Alan Coopersmith <alan.coopersmith@oracle.com>