Xvproto.h: fix xvQueryEncodingsReply to match actual wire protocol
padl3 was declared as CARD32, which for proper alignment, got two
additional bytes of padding silently inserted before it.  Declaring
it as CARD16 aligns it properly after the CARD16 num_encodings.

Fixes clang warning:
Xvproto.h:462:10: warning: padding struct 'struct _QueryEncodingsReply' with 2 bytes to align 'padl3' [-Wpadded]
  CARD32 padl3 B32;
         ^
This makes sizeof(xvQueryEncodingsReply) match the sz_xvQueryEncodingsReply
which is used in libXv & the X server for the size sent over the wire.

Reported-by: Daniel Martin <consume.noise@gmail.com>
Reviewed-by: Julien Cristau <jcristau@debian.org>
Signed-off-by: Alan Coopersmith <alan.coopersmith@oracle.com>