Add comments about sizes of dynamically allocated arrays
Many arrays have sz_* or size_* fields to list the number of entries
allocated, and num_* fields to record the number of entries used.

Others use num_* for the number allocated, or size based on max_key_code.

And a few are just plain trying to mess with your head.  (I'm looking at
you XkbNamesRec & XkbKeyTypeRec.)

It sure would have been nice if all the XKB authors could have picked
a single convention and stuck to it, but we're 20 years too late for
that now, so just document it so I can stop reverse engineering from
the code every time I need to see if we got a bounds check right or not.

Signed-off-by: Alan Coopersmith <alan.coopersmith@oracle.com>