HideCursor/ShowCursor: Fix the interpretation of the window argument
In the implementation of HideCursor/ShowCursor in the xserver code,
the window passed in only identifies a screen, not a portion of the
window tree, and it's been that way since the beginning. Changing
the implementation at this point would likely break existing code,
and the main usage of this protocol (magnifiers) only needs the
global behavior, so change the protocol to document the actual
behavior.

Reviewed-by: Keith Packard <keithp@keithp.com>
Reviewed-by: Daniel Stone <daniels@collabora.com>
Signed-off-by: Owen W. Taylor <otaylor@fishsoup.net>