Move the headers under include/X11/...
I kinda hate to do this, it was nice to have everything in one place.
However, the meson build wants to be able to wrap this module as a
dependency, and code that depends on these headers includes them in the
form:

    #include <X11/Xfuncproto.h>

As a result, any include path meson can construct needs to point to the
root of a hierarchy that has the same path layout as an installed copy,
hence this change.

Signed-off-by: Adam Jackson <ajax@redhat.com>