randr: Add Leases. [v4]
A "lease" is a set of crtc and output resources granted to another
application for use outside of X. These will not be usable through the
X protocol until the lease terminates. Leased outputs will be seen as
disconnected, leased CRTCs will be seen as not usable with any output.

v2:
	Delete output grabs
	Add LeaseNotify events
	Add FreeLease with option to terminate

v3:
	Clarify a couple of lease behaviors:

	* You can lease an in-use object, it makes the X server stop
	  using it, you don't get an error back.

	* There's no explicit 'Disabled' state for a crtc, when a crtc
	  is disabled, it just has a set of reported values for
	  GetCrtcInfo.

v4:
	Integrate into merged xorgproto repo

Signed-off-by: Keith Packard <keithp@keithp.com>