Combine *_EXTRA_DIST into EXTRA_DIST in Makefile.am
automake only recognizes EXTRA_DIST - but it looked like it worked since
the *.pc.in were included since they're listed in configure.ac, unless
you tried building from the tarballs with --enable-legacy, and then it
broke.

Also adds --enable-legacy to the flags that "make distcheck" uses to
configure, to catch problems like this.

Fixes: https://gitlab.freedesktop.org/xorg/proto/xorgproto/issues/12

Signed-off-by: Alan Coopersmith <alan.coopersmith@oracle.com>