keysymdef: Fix left/rightanglebracket Unicode equivalent
As of version 7.7, the X protocol does not define a Unicode equivalent for
them.  The U+27E8 and U+27E9 equivalents were introduced by 618956f1f ("The
big keysym cleanup, to bring implementation in line with the recent revision
of Appendix A of the protocol spec."), but as xterm Patch #226 explicitly
notes, U+2329 and U+232A should be used rather than U+27E8 and U+27E9.  Gtk
also inherited this.

Signed-off-by: Samuel Thibault <samuel.thibault@ens-lyon.org>