Integrate the keysym verifier into make check
autotools can't pass arguments, so let's default to 'verify' in the script
itself and for distcheck to succeed, we need to set an environment variable to
search for the header (it's an out-of-tree build).

And due to the very faint chance of there being no python during the
xorgproto build, let's make that conditional too.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>