gitlab CI: add a job to compare meson and autotools standard installs
Build and install with meson, build and install with autotools and then run
diff to compare the two directory trees. They should be the same.

This does not install the legacy protocols, they're behind a configure switch.
The spec-build is disabled in autotools because we know meson doesn't do that
yet, so no point in comparing those.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>